Struct cld2::LanguageScore

pub struct LanguageScore {
    pub language: Option<Lang>,
    pub percent: u8,
    pub normalized_score: f64,
}

Detailed information about how well the input text matched a specific language.

Fields

language

The language matched.

percent

The percentage of the text which appears to be in this language. Between 0 and 100.

normalized_score

Scores near 1.0 indicate a "normal" text for this language. Scores further away from 1.0 indicate strange or atypical texts.